requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
docstring_parser==0.17.0
executing==2.2.1
fhirclient==4.1.0
google-api-core==2.28.0
google-auth==2.42.0
google-cloud-aiplatform==1.127.0
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.3
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-genai==1.50.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jsonschema==4.20.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
lxml==6.0.2
matplotlib-inline==0.2.1
more-itertools==10.8.0
nest-asyncio==1.6.0
numpy==1.26.2
packaging==25.0
pandas==2.1.4
parso==0.8.5
pendulum==3.1.0
pexpect==4.9.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.1
pyzmq==27.1.0
referencing==0.37.0
requests==2.31.0
requests-file==3.0.1
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
shapely==2.1.2
simple-salesforce==1.12.5
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.2
tqdm==4.66.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
zeep==4.3.2
